+++ resources/views/articles/show.blade.php
@@ -22,8 +22,10 @@
     <div class="card-body">
         <h5 class="card-title">{{$comment->title}} </h5> <h4 class="card-title">({{$comment->created_at}})</h4>
         <p class="card-text">{{$comment->text}}</p>
-        <a href="/comment/{{$comment->id}}/edit" class="btn btn-secondary">Редактировать</a>
-        <a href="/comment/{{$comment->id}}/delete" class="btn btn-secondary">Удалить</a>
+        @can('update-comment', $comment)
+            <a href="/comment/{{$comment->id}}/edit" class="btn btn-secondary">Редактировать</a>
+            <a href="/comment/{{$comment->id}}/delete" class="btn btn-secondary">Удалить</a>
+        @endcan
     </div>
 @endforeach
 </div>
